infracloud_project/requirements.txt
asgiref==3.4.1
backcall==0.2.0
decorator==5.0.9
Django==3.2.5
djangorestframework==3.12.4
ipykernel==4.10.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.2.0
jupyter-core==4.7.1
nest-asyncio==1.5.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==1.0.18
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.1.0
six==1.16.0
sqlparse==0.4.1
tornado==5.1.1
traitlets==4.3.3
typing-extensions==3.10.0.0
wcwidth==0.2.5
